requirements.txt
"fastapi[standard]"
passlib
pydantic
PyJWT
python-dotenv
SQLAlchemy
